+++ Changelog.md
@@ -36,6 +36,7 @@
 - Fix: Disguisable blocks that are disguised as blocks that emit light (e.g. Glowstone) don't emit light
 - Fix: The Motion Activated Light is constantly turned on when placed near a Sentry or armor stand
 - Fix: Prohibited items get removed from an Inventory Scanner if a Storage Module is added/removed
+- Fix: Changing the owner of an Inventory Scanner doesn't change the owner of the corresponding fields
 - Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Removed: "allowCodebreakerItem" config option. Disabling the Codebreaker is now achieveable by setting "codebreaker_chance" to a negative value
 - Misc.: More texture and model changes